I hate IDE

+++ arch/x86_64/include/ide.h
@@ -121,7 +121,7 @@ typedef struct {
     uint32_t lba;
     uint8_t sectors;
     uint16_t selector;
-    uint32_t buffer;
+    uint64_t buffer;
 } ide_rw_t;
 
 extern ide_channel_reg_t __ide_channels[2];
